--- components/cards/attributes.json
@@ -1,7 +1,0 @@
-[
-  {},
-  { "name": "Sp", "color": "red", "fullname": "Sparkle" },
-  { "name": "Br", "color": "lightblue", "fullname": "Brilliant" },
-  { "name": "Gl", "color": "green", "fullname": "Glitter" },
-  { "name": "Fl", "color": "yellow", "fullname": "Flash" }
-]
