+++ data/liveSkill.json
@@ -1,0 +1,188 @@
+{
+  "5": {
+    "1": [
+      [5, 5],
+      [5, 6],
+      [5, 7],
+      [5, 8],
+      [5, 10],
+      [5, 11],
+      [5, 12],
+      [5, 15],
+      [5, 20],
+      [5, 25]
+    ],
+    "2": [
+      [5, 5],
+      [5, 6],
+      [5, 7],
+      [5, 8],
+      [5, 10],
+      [5, 11],
+      [5, 12],
+      [5, 15],
+      [5, 20],
+      [5, 25]
+    ],
+    "3": [
+      [5, 10],
+      [5, 12],
+      [5, 15],
+      [5, 17],
+      [5, 20],
+      [5, 22],
+      [5, 25],
+      [5, 30],
+      [5, 40],
+      [5, 50]
+    ],
+    "4": [
+      [5, 20],
+      [5, 25],
+      [5, 30],
+      [5, 35],
+      [5, 40],
+      [5, 45],
+      [5, 50],
+      [5, 60],
+      [5, 70],
+      [5, 80]
+    ],
+    "5": [
+      [5, 20],
+      [5, 25],
+      [5, 30],
+      [5, 35],
+      [5, 40],
+      [5, 45],
+      [5, 50],
+      [5, 60],
+      [5, 70],
+      [5, 80]
+    ]
+  },
+  "8": {
+    "1": [
+      [8, 3],
+      [8, 4],
+      [8, 5],
+      [8, 6],
+      [8, 7],
+      [8, 8],
+      [8, 9],
+      [8, 10],
+      [8, 12],
+      [8, 15]
+    ],
+    "2": [
+      [8, 3],
+      [8, 4],
+      [8, 5],
+      [8, 6],
+      [8, 7],
+      [8, 8],
+      [8, 9],
+      [8, 10],
+      [8, 12],
+      [8, 15]
+    ],
+    "3": [
+      [8, 6],
+      [8, 8],
+      [8, 10],
+      [8, 12],
+      [8, 14],
+      [8, 16],
+      [8, 18],
+      [8, 20],
+      [8, 25],
+      [8, 30]
+    ],
+    "4": [
+      [8, 13],
+      [8, 16],
+      [8, 20],
+      [8, 23],
+      [8, 26],
+      [8, 30],
+      [8, 35],
+      [8, 40],
+      [8, 45],
+      [8, 50]
+    ],
+    "5": [
+      [8, 13],
+      [8, 16],
+      [8, 20],
+      [8, 23],
+      [8, 26],
+      [8, 30],
+      [8, 35],
+      [8, 40],
+      [8, 45],
+      [8, 50]
+    ]
+  },
+  "12": {
+    "1": [
+      [12, 1],
+      [12, 2],
+      [12, 3],
+      [12, 4],
+      [12, 5],
+      [12, 6],
+      [12, 7],
+      [12, 8],
+      [12, 9],
+      [12, 10]
+    ],
+    "2": [
+      [12, 1],
+      [12, 2],
+      [12, 3],
+      [12, 4],
+      [12, 5],
+      [12, 6],
+      [12, 7],
+      [12, 8],
+      [12, 9],
+      [12, 10]
+    ],
+    "3": [
+      [12, 4],
+      [12, 5],
+      [12, 6],
+      [12, 7],
+      [12, 8],
+      [12, 9],
+      [12, 10],
+      [12, 12],
+      [12, 15],
+      [12, 20]
+    ],
+    "4": [
+      [12, 8],
+      [12, 10],
+      [12, 12],
+      [12, 14],
+      [12, 16],
+      [12, 18],
+      [12, 20],
+      [12, 25],
+      [12, 30],
+      [12, 35]
+    ],
+    "5": [
+      [12, 8],
+      [12, 10],
+      [12, 12],
+      [12, 14],
+      [12, 16],
+      [12, 18],
+      [12, 20],
+      [12, 25],
+      [12, 30],
+      [12, 35]
+    ]
+  }
+}
