+++ data/substatAttribute.json
@@ -1,0 +1,1 @@
+["Da", "Vo", "Pf"]
