+++ data/supportSkill.json
@@ -1,0 +1,119 @@
+[
+  {
+    "id": 1,
+    "daydream_name": "VoltageSupport",
+    "1": [[2], [3], [5], [8], [10]],
+    "2": [[2], [3], [5], [8], [10]],
+    "3": [[2], [3], [5], [8], [10]],
+    "4": [[5], [8], [10], [15], [20]],
+    "5": [[5], [8], [10], [15], [20]]
+  },
+  {
+    "id": 2,
+    "daydream_name": "ComboSupport",
+    "5": [
+      [1, 0, 3],
+      [1, 1, 3],
+      [1, 1, 4],
+      [2, 1, 4],
+      [2, 1, 5]
+    ]
+  },
+  {
+    "id": 3,
+    "daydream_name": "PerfectSupport",
+    "5": [[1], [2], [3], [4], [5]]
+  },
+  {
+    "id": 4,
+    "daydream_name": "EnsembleTimeSupport",
+    "1": [[2], [3], [5], [8], [10]],
+    "2": [[2], [3], [5], [8], [10]],
+    "3": [[2], [3], [5], [8], [10]],
+    "4": [[5], [8], [10], [15], [20]],
+    "5": [[5], [8], [10], [15], [20]]
+  },
+  {
+    "id": 5,
+    "daydream_name": "PieceUpDaS",
+    "5": [[50], [60], [70], [80], [100]],
+    "4": [[25], [35], [50], [65], [75]],
+    "3": [[10], [20], [30], [40], [50]]
+  },
+  {
+    "id": 6,
+    "daydream_name": "PieceUpDaM",
+    "5": [[50], [60], [70], [80], [100]],
+    "4": [[25], [35], [50], [65], [75]],
+    "3": [[10], [20], [30], [40], [50]]
+  },
+  {
+    "id": 7,
+    "daydream_name": "PieceUpDaL",
+    "5": [[50], [60], [70], [80], [100]],
+    "4": [[25], [35], [50], [65], [75]],
+    "3": [[10], [20], [30], [40], [50]]
+  },
+  {
+    "id": 8,
+    "daydream_name": "PieceUpDaAll",
+    "5": [[25], [35], [50], [65], [75]]
+  },
+  {
+    "id": 9,
+    "daydream_name": "PieceUpVoS",
+    "5": [[50], [60], [70], [80], [100]],
+    "4": [[25], [35], [50], [65], [75]],
+    "3": [[10], [20], [30], [40], [50]]
+  },
+  {
+    "id": 10,
+    "daydream_name": "PieceUpVoM",
+    "5": [[50], [60], [70], [80], [100]],
+    "4": [[25], [35], [50], [65], [75]],
+    "3": [[10], [20], [30], [40], [50]]
+  },
+  {
+    "id": 11,
+    "daydream_name": "PieceUpVoL",
+    "5": [[50], [60], [70], [80], [100]],
+    "4": [[25], [35], [50], [65], [75]],
+    "3": [[10], [20], [30], [40], [50]]
+  },
+  {
+    "id": 12,
+    "daydream_name": "PieceUpVoAll",
+    "5": [[25], [35], [50], [65], [75]]
+  },
+  {
+    "id": 13,
+    "daydream_name": "PieceUpPfS",
+    "5": [[50], [60], [70], [80], [100]],
+    "4": [[25], [35], [50], [65], [75]],
+    "3": [[10], [20], [30], [40], [50]]
+  },
+  {
+    "id": 14,
+    "daydream_name": "PieceUpPfM",
+    "5": [[50], [60], [70], [80], [100]],
+    "4": [[25], [35], [50], [65], [75]],
+    "3": [[10], [20], [30], [40], [50]]
+  },
+  {
+    "id": 15,
+    "daydream_name": "PieceUpPfL",
+    "5": [[50], [60], [70], [80], [100]],
+    "4": [[25], [35], [50], [65], [75]],
+    "3": [[10], [20], [30], [40], [50]]
+  },
+  {
+    "id": 16,
+    "daydream_name": "PieceUpPfAll",
+    "5": [[25], [35], [50], [65], [75]]
+  },
+  {
+    "id": 28,
+    "daydream_name": "PieceUpAllAll",
+    "5": [[25], [35], [50], [65], [75]]
+  }
+]
