+++ package.json
@@ -29,7 +29,8 @@
     "cldrjs": "^0.5.5",
     "cookies-next": "^2.0.4",
     "country-flag-icons": "^1.4.26",
-    "firebase": "^9.7.0",
+    "firebase": "^9.9.0",
+    "firebase-admin": "^11.0.0",
     "framer-motion": "^6.3.3",
     "globalize": "^1.7.0",
     "iana-tz-data": "^2019.1.0",
@@ -37,13 +38,14 @@
     "lodash.debounce": "^4.0.8",
     "lodash.throttle": "^4.1.1",
     "multiselect-react-dropdown": "^2.0.21",
-    "next": "^12.1.5",
+    "next": "^12.2.2",
+    "next-firebase-auth": "^1.0.0-canary.9",
     "next-i18next": "^11.0.0",
     "normalize.css": "^8.0.1",
     "numbro": "^2.3.6",
     "numeral": "^2.0.6",
-    "react": "^18.1.0",
-    "react-dom": "^18.1.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-feather": "^2.0.9",
     "react-firebase-hooks": "^5.0.3",
     "react-globalize": "^1.0.0",
@@ -81,7 +83,7 @@
     "@iconify/icons-tabler": "^1.2.8",
     "@iconify/react": "^3.2.0",
     "babel-plugin-styled-components": "^2.0.2",
-    "eslint": "^7.32.0",
+    "eslint": "^8.19.0",
     "eslint-config-airbnb": "^18.2.1",
     "eslint-config-next": "^12.0.4",
     "eslint-config-prettier": "^8.3.0",
