+++ styles/CharactersPageImageCard.module.scss
